# optimize regular expressions