@mvdan 2022-01-21
I'll send a follow-up PR shortly; FYI in case that could cause conflicts.